+++ videoplayer-api/src/main/java/intbird/soft/lib/video/player/api/bean/MediaFileType.kt
@@ -1,5 +1,10 @@
 package intbird.soft.lib.video.player.api.bean
 
+/**
+ * created by intbird
+ * on 2020/9/8
+ * DingTalk id: intbird
+ */
 enum class MediaFileType(val type: String) {
     FILE("file:///"),
     HTTP("http://"),