+++ videoplayer/src/main/java/intbird/soft/lib/video/player/main/dialog/SingleChooseCallback.kt
@@ -9,5 +9,5 @@ import intbird.soft.lib.video.player.api.bean.MediaCheckedData
  */
 
 interface SingleChooseCallback {
-    fun onChooseItem(mediaCheckedData: MediaCheckedData)
+    fun onChooseItem(index: Int, mediaCheckedData: MediaCheckedData)
 }