+++ videoplayer/src/main/java/intbird/soft/lib/video/player/main/dialog/type/SingleChooseType.kt
@@ -1,0 +1,10 @@
+package intbird.soft.lib.video.player.main.dialog.type
+
+/**
+ * created by Bird
+ * on 2020/9/11
+ * DingTalk id: intbird
+ */
+enum class SingleChooseType {
+    NONE, CLARITY, RATES, TEXT
+}