+++ videoplayer/src/main/java/intbird/soft/lib/video/player/main/player/player/delegate/PlayerDelegate.kt
@@ -1,0 +1,14 @@
+package intbird.soft.lib.video.player.main.player.player.delegate
+
+/**
+ * created by intbird
+ * on 2020/9/10
+ * DingTalk id: intbird
+ *
+ * last and next all in items,so the player doest not impl
+ */
+interface PlayerDelegate {
+    fun delegatePlay(): Boolean
+    fun delegateLast(): Boolean
+    fun delegateNext(): Boolean
+}